+++ src/config.ts
@@ -1,7 +1,7 @@
 /*
  * Centralized location for non-user configurable settings.
  */
-import { CacheKey, NonScoreWeightName, TagTootsType, type ApiCacheKey } from "./enums";
+import { CacheKey, NonScoreWeightName, TagTootsCategory, type ApiCacheKey } from "./enums";
 import { isDebugMode, isLoadTest, isQuickMode } from "./helpers/environment_helpers";
 import { type NonScoreWeightInfoDict } from "./types";
 
@@ -255,13 +255,13 @@ class Config implements ConfigType {
                 initialMaxRecords: MAX_ENDPOINT_RECORDS_TO_PULL,
                 minutesUntilStale: 4 * MINUTES_IN_HOUR,
             },
-            [TagTootsType.FAVOURITED]: {
+            [TagTootsCategory.FAVOURITED]: {
                 minutesUntilStale: 60,
             },
-            [TagTootsType.PARTICIPATED]: {
+            [TagTootsCategory.PARTICIPATED]: {
                 minutesUntilStale: 20,
             },
-            [TagTootsType.TRENDING]: {
+            [TagTootsCategory.TRENDING]: {
                 minutesUntilStale: 15,
             },
         } as ApiDataConfig,