+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Rename TagTootsType enum to TagTootsCategory
 
 ### v0.74.1
 * Only do the followed hashtag additional scan in finishFeedUpdate(), not for each fetch because it's slow